feat: refactor seq_io specific code into its own module
